Upgraded implementation to use KnexJS
To help facilitate using multiple databases PostgreSQL, MySQL, MSSQL, and Oracle I've decided to use KnexJS so the code can be written in a database agnostic way.  If this turns out to be the wrong library we may switch to a different one.

In addition to implementing Knex I also exctracted the quick browse data tab to be a separate Vue component, not so much for reuse as it is for making the code more manageable.  I also added buttons for navigating the quick browse pagination.